requirements.txt
streamlit
google-genai
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pillow
streamlit-paste-button
openai
st-copy-to-clipboard
anthropic
pandas
watchdog
google-auth
